+++ helper/docker_cri_adapter.go
@@ -51,7 +51,6 @@ var criRuntimeWrapper *CRIRuntimeWrapper
 
 // CRIRuntimeWrapper wrapper for containerd client
 type CRIRuntimeWrapper struct {
-	// client       *containerd.Client
 	dockerCenter *DockerCenter
 
 	client         cri.RuntimeServiceClient
@@ -216,16 +215,64 @@ func NewCRIRuntimeWrapper(dockerCenter *DockerCenter) (*CRIRuntimeWrapper, error
 	}, nil
 }
 
+func (cw *CRIRuntimeWrapper) fetchOne(containerID string) error {
+	dockerContainer, sandboxID, status, err := cw.createContainerInfo(containerID)
+	if err != nil {
+		return err
+	}
+	cw.wrapperK8sInfoById(sandboxID, dockerContainer)
+
+	if logger.DebugFlag() {
+		bytes, _ := json.Marshal(dockerContainer)
+		logger.Debugf(context.Background(), "cri crate container info : %s", string(bytes))
+	}
+
+	cw.dockerCenter.updateContainer(containerID, dockerContainer)
+	cw.containersLock.Lock()
+	cw.containers[containerID] = status
+	cw.containersLock.Unlock()
+	return nil
+}
+
+func (cw *CRIRuntimeWrapper) wrapperK8sInfoById(sandboxID string, detail *DockerInfoDetail) {
+	ctx, cancel := getContextWithTimeout(time.Second * 10)
+	status, err := cw.client.PodSandboxStatus(ctx, &cri.PodSandboxStatusRequest{
+		PodSandboxId: sandboxID,
+		Verbose:      true,
+	})
+	cancel()
+	if err != nil {
+		logger.Debug(context.Background(), "fetchone cannot read k8s info from sandbox, sandboxID", sandboxID)
+		return
+	}
+	cw.wrapperK8sInfoByLabels(status.GetStatus().GetLabels(), detail)
+}
+
+func (cw *CRIRuntimeWrapper) wrapperK8sInfoByLabels(sandboxLabels map[string]string, detail *DockerInfoDetail) {
+	if detail.K8SInfo == nil || sandboxLabels == nil {
+		return
+	}
+	if detail.K8SInfo.Labels == nil {
+		detail.K8SInfo.Labels = make(map[string]string)
+	}
+	for k, v := range sandboxLabels {
+		if strings.HasPrefix(k, k8sInnerLabelPrefix) || strings.HasPrefix(k, k8sInnerAnnotationPrefix) {
+			continue
+		}
+		detail.K8SInfo.Labels[k] = v
+	}
+}
+
 // createContainerInfo convert cri container to docker spec to adapt the history logic.
-func (cw *CRIRuntimeWrapper) createContainerInfo(_ context.Context, c *cri.Container) (*DockerInfoDetail, error) {
+func (cw *CRIRuntimeWrapper) createContainerInfo(containerID string) (detail *DockerInfoDetail, sandboxID string, state cri.ContainerState, err error) {
 	ctx, cancel := getContextWithTimeout(time.Second * 10)
 	status, err := cw.client.ContainerStatus(ctx, &cri.ContainerStatusRequest{
-		ContainerId: c.GetId(),
+		ContainerId: containerID,
 		Verbose:     true,
 	})
 	cancel()
 	if err != nil {
-		return nil, err
+		return nil, "", cri.ContainerState_CONTAINER_UNKNOWN, err
 	}
 
 	var ci containerdcriserver.ContainerInfo
@@ -235,15 +282,15 @@ func (cw *CRIRuntimeWrapper) createContainerInfo(_ context.Context, c *cri.Conta
 			foundInfo = true
 			ci, err = parseContainerInfo(info)
 			if err != nil {
-				logger.Errorf(context.Background(), "CREATE_CONTAINERD_INFO_ALARM", "failed to parse container info, containerId: %s, data: %s, error: %v", c.GetId(), info, err)
+				logger.Errorf(context.Background(), "CREATE_CONTAINERD_INFO_ALARM", "failed to parse container info, containerId: %s, data: %s, error: %v", containerID, info, err)
 			}
 		}
 	}
 	if !foundInfo {
-		logger.Warningf(context.Background(), "CREATE_CONTAINERD_INFO_ALARM", "can not find container info from CRI::ContainerStatus, containerId: %s", c.GetId())
+		logger.Warningf(context.Background(), "CREATE_CONTAINERD_INFO_ALARM", "can not find container info from CRI::ContainerStatus, containerId: %s", containerID)
 	}
 
-	labels := c.GetLabels()
+	labels := status.GetStatus().GetLabels()
 	if labels == nil {
 		labels = map[string]string{}
 	}
@@ -254,7 +301,7 @@ func (cw *CRIRuntimeWrapper) createContainerInfo(_ context.Context, c *cri.Conta
 	}
 
 	dockerContainer := &docker.Container{
-		ID:      c.GetId(),
+		ID:      containerID,
 		LogPath: status.GetStatus().GetLogPath(),
 		Config: &docker.Config{
 			Labels: labels,
@@ -269,8 +316,8 @@ func (cw *CRIRuntimeWrapper) createContainerInfo(_ context.Context, c *cri.Conta
 		},
 	}
 
-	if c.GetMetadata() != nil {
-		dockerContainer.Name = c.GetMetadata().GetName()
+	if status.GetStatus().GetMetadata() != nil {
+		dockerContainer.Name = status.GetStatus().GetMetadata().GetName()
 	}
 
 	if ci.RuntimeSpec != nil && ci.RuntimeSpec.Process != nil {
@@ -309,7 +356,7 @@ func (cw *CRIRuntimeWrapper) createContainerInfo(_ context.Context, c *cri.Conta
 	dockerContainer.HostnamePath = hostnamePath
 	dockerContainer.HostsPath = hostsPath
 
-	return cw.dockerCenter.CreateInfoDetail(dockerContainer, envConfigPrefix, false), nil
+	return cw.dockerCenter.CreateInfoDetail(dockerContainer, envConfigPrefix, false), ci.SandboxID, status.GetStatus().GetState(), nil
 }
 
 func (cw *CRIRuntimeWrapper) fetchAll() error {
@@ -331,11 +378,14 @@ func (cw *CRIRuntimeWrapper) fetchAll() error {
 	}
 
 	containerMap := make(map[string]*DockerInfoDetail)
+	cw.containersLock.Lock()
+	defer cw.containersLock.Unlock()
 	for _, container := range containersResp.Containers {
 		if container.State == cri.ContainerState_CONTAINER_EXITED || container.State == cri.ContainerState_CONTAINER_UNKNOWN {
 			continue
 		}
-		dockerContainer, err := cw.createContainerInfo(ctx, container)
+		cw.containers[container.GetId()] = container.State
+		dockerContainer, _, _, err := cw.createContainerInfo(container.GetId())
 		if err != nil {
 			logger.Debug(context.Background(), "Create container info from cri-runtime error", err)
 			continue
@@ -343,20 +393,18 @@ func (cw *CRIRuntimeWrapper) fetchAll() error {
 		containerMap[container.GetId()] = dockerContainer
 
 		// append the pod labels to the k8s info.
-		if sandbox, ok := sandboxMap[container.PodSandboxId]; ok && dockerContainer.K8SInfo != nil {
-			if dockerContainer.K8SInfo.Labels == nil {
-				dockerContainer.K8SInfo.Labels = make(map[string]string)
-			}
-			for k, v := range sandbox.Labels {
-				if strings.HasPrefix(k, k8sInnerLabelPrefix) || strings.HasPrefix(k, k8sInnerAnnotationPrefix) {
-					continue
-				}
-				dockerContainer.K8SInfo.Labels[k] = v
-			}
+		if sandbox, ok := sandboxMap[container.PodSandboxId]; ok {
+			cw.wrapperK8sInfoByLabels(sandbox.GetLabels(), dockerContainer)
 		}
 		logger.Debug(context.Background(), "Create container info from cri-runtime success, info", *dockerContainer.ContainerInfo, "config", *dockerContainer.ContainerInfo.Config, "detail", *dockerContainer)
 	}
 	cw.dockerCenter.updateContainers(containerMap)
+
+	for k := range cw.containers {
+		if _, ok := containerMap[k]; !ok {
+			delete(cw.containers, k)
+		}
+	}
 	return nil
 }
 
@@ -384,45 +432,46 @@ func (cw *CRIRuntimeWrapper) loopSyncContainers() {
 func (cw *CRIRuntimeWrapper) syncContainers() error {
 	ctx, cancel := getContextWithTimeout(time.Second * 20)
 	defer cancel()
+	logger.Debug(context.Background(), "cri sync containers", "begin")
 	containersResp, err := cw.client.ListContainers(ctx, &cri.ListContainersRequest{})
 	if err != nil {
 		return err
 	}
 
-	cw.containersLock.Lock()
-	defer cw.containersLock.Unlock()
-
-	oldContainers := cw.containers
-
 	newContainers := map[string]*cri.Container{}
-	for i := range containersResp.Containers {
+	for i, container := range containersResp.Containers {
+		if container.State == cri.ContainerState_CONTAINER_EXITED || container.State == cri.ContainerState_CONTAINER_UNKNOWN {
+			continue
+		}
 		id := containersResp.Containers[i].GetId()
 		newContainers[id] = containersResp.Containers[i]
 	}
 
 	// update container
 	for id, container := range newContainers {
-		if oldState, ok := oldContainers[id]; !ok || oldState != container.State {
-			ctx, cancel := getContextWithTimeout(time.Second * 10)
-			dockerContainer, err := cw.createContainerInfo(ctx, container)
-			cancel()
-			if err != nil {
+		cw.containersLock.RLock()
+		oldState, ok := cw.containers[id]
+		cw.containersLock.RUnlock()
+		if !ok || oldState != container.State {
+			logger.Debug(context.Background(), "cri sync containers fetchOne", id)
+			if err := cw.fetchOne(id); err != nil {
 				logger.Errorf(context.Background(), "CREATE_CONTAINERD_INFO_ALARM", "failed to createContainerInfo, containerId: %s, error: %v", id, err)
 				continue
 			}
-			cw.dockerCenter.updateContainer(id, dockerContainer)
-			cw.containers[id] = container.State
 		}
 	}
 
 	// delete container
+	cw.containersLock.Lock()
+	defer cw.containersLock.Unlock()
 	for oldID := range cw.containers {
 		if _, ok := newContainers[oldID]; !ok {
 			cw.dockerCenter.markRemove(oldID)
+			logger.Debug(context.Background(), "cri sync containers remove", oldID)
 			delete(cw.containers, oldID)
 		}
 	}
-
+	logger.Debug(context.Background(), "cri sync containers", "done")
 	return nil
 }
 
@@ -442,38 +491,6 @@ func (cw *CRIRuntimeWrapper) sweepCache() {
 		}
 	}
 	cw.rootfsLock.Unlock()
-}
-
-func (cw *CRIRuntimeWrapper) run() error {
-	logger.Init()
-	logger.Info(context.Background(), "CRIRuntime background syncer", "start")
-	_ = cw.fetchAll()
-	logger.Info(context.Background(), "CRIRuntime background syncer", "gogogogo")
-
-	timerFetch := func() {
-		defer dockerCenterRecover()
-		lastFetchAllTime := time.Now()
-		for {
-			time.Sleep(time.Duration(10) * time.Second)
-			logger.Debug(context.Background(), "docker clean timeout container info", "start")
-			cw.dockerCenter.cleanTimeoutContainer()
-			logger.Debug(context.Background(), "docker clean timeout container info", "done")
-			if time.Since(lastFetchAllTime) >= FetchAllInterval {
-				logger.Info(context.Background(), "CRIRuntime fetch all", "start")
-				cw.dockerCenter.readStaticConfig(true)
-				err := cw.fetchAll()
-				logger.Info(context.Background(), "CRIRuntime fetch all", err)
-				lastFetchAllTime = time.Now()
-				cw.sweepCache()
-			}
-
-		}
-	}
-	go timerFetch()
-
-	go cw.loopSyncContainers()
-
-	return nil
 }
 
 func getContextWithTimeout(timeout time.Duration) (context.Context, context.CancelFunc) {